requirements.txt
tensorflow
pandas
pytorch
torchvision